RequirementMachine: Move some code to a new PropertyRelations.cpp file
